% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSimplexes.R
\name{analyze}
\alias{analyze}
\title{Analyze tessellation}
\usage{
analyze(tes, func, ...)
}
\arguments{
\item{tes}{The tessellation object.}

\item{func}{The function to use to analyze the critical points of each simplex.}

\item{...}{Additional arguments required for func}
}
\value{
Vector of boolean specifying if all critical points in the each simplex pass the 
criterion evaluated in func.
}
\description{
Analyze the tessellation of the target class
}
\keyword{internal}
