% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalPoints.R
\name{createCriticalPointMap}
\alias{createCriticalPointMap}
\title{Find critical points map}
\usage{
createCriticalPointMap(tes, method, tes1_index, tes2_index)
}
\arguments{
\item{tes}{The tessellation object.}

\item{method}{The method to use.}
}
\value{
Row matrix of critical points
}
\description{
Find critical points map: The points that will be checked within a polytope.
Method 1 (Necessary and Sufficient) - Complexity : N * D^3
Method 2 (Sufficient) - Complexity : N * D
}
\keyword{internal}
