% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{density}
\alias{density}
\title{¨Get the density for a given class at given points}
\usage{
density(tes, index, points)
}
\arguments{
\item{tes}{The tessellations object}

\item{index}{The class index}

\item{points}{The points}
}
\value{
Numeric vector giving value of density at each point.
}
\description{
¨Get the density for a given class at given points
}
