% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{densityplot}
\alias{densityplot}
\title{Plot the tessellation densities.}
\usage{
densityplot(tes, alpha = 1)
}
\arguments{
\item{tes}{The tessellations object}

\item{alpha}{The alpha parameter for the density shading.}
}
\description{
Plot the tessellation densities.
}
\keyword{internal}
