% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSimplexes.R
\name{dom}
\alias{dom}
\title{Function that checks for dominance}
\usage{
dom(pnts, tes)
}
\arguments{
\item{tes}{The tessellations object}

\item{pnt}{The point(s) to test at}
}
\value{
Boolean vector specifying whether the target class's probability is the maximum at
each point.
}
\description{
Function that checks for dominance. Passed to analyzeTessellation.
}
