% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalPoints.R
\name{findCriticalPointsForPolytope}
\alias{findCriticalPointsForPolytope}
\title{Find Critical Points For Polytope}
\usage{
findCriticalPointsForPolytope(tri, tes1_index, tes2_index, tes, method)
}
\arguments{
\item{tri}{The triangle to find critical points for.}

\item{tes}{The tessellation object.}

\item{method}{The method to use.}
}
\value{
Row matrix of critical points
}
\description{
Find Critical Points For Polytope
}
\keyword{internal}
