% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalPoints.R
\name{findExitPoints}
\alias{findExitPoints}
\title{Complexity (WC): N * D^3}
\usage{
findExitPoints(enclosed_indices, tri, tes1_index, tes2_index, tes, col)
}
\value{
Matrix of exit points
}
\description{
Complexity (WC): N * D^3
}
