% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalPoints.R
\name{findExitPointsForVertex}
\alias{findExitPointsForVertex}
\title{Complexity (WC): N * D^3}
\usage{
findExitPointsForVertex(enclosed_index, enclosed_indices, tri1, tes1_index,
  tes2_index, tes, col)
}
\value{
Matrix of exit points
}
\description{
Complexity (WC): N * D^3
}
