% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalPoints.R
\name{findExitPointsInPolytope}
\alias{findExitPointsInPolytope}
\title{Complexity: D^3}
\usage{
findExitPointsInPolytope(tri2, vertex_index, enclosed_vertices, tri1, tes1,
  tes2, tes, col)
}
\value{
Matrix of exit points
}
\description{
Complexity: D^3
}
