% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalPoints.R
\name{findExitPointsInPolytopeOfEdge}
\alias{findExitPointsInPolytopeOfEdge}
\title{Complexity: D^2}
\usage{
findExitPointsInPolytopeOfEdge(other_vertex, vertex, enclosed_vertices, tri1,
  tes1_index, classdata1, classdata2, tes)
}
\value{
Point of exit or NULL if no such point
}
\description{
Complexity: D^2
}
