% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIntersect.R
\name{findIntersect}
\alias{findIntersect}
\title{Find if a line intersects a convex polytope}
\usage{
findIntersect(p0, p1, vertices, normalvectors)
}
\arguments{
\item{p0}{The start of the line}

\item{p1}{The end of the line}

\item{vertices}{A matrix, each column giving a vertex for each face, 
ordered as normalvectors}

\item{normalvectors}{A matrix, each column giving a normal vector for each face, 
ordered as vertices}
}
\value{
FALSE if no intersection found. List of entry and exit points if interset is
found.
}
\description{
Find if a line intersects a convex polytope
}
\keyword{internal}
