% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{getBasicData}
\alias{getBasicData}
\title{Get a small set of data}
\usage{
getBasicData(a = 4, b = 6, c = 8)
}
\arguments{
\item{a}{An}

\item{b}{Bn}

\item{c}{Cn}
}
\value{
A data matrix
}
\description{
Get a small set of data. This is a wrapper for getTestData using the first three arguments.
}
