% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalVector.R
\name{getHPCoefs}
\alias{getHPCoefs}
\title{Get the hyperplane coefficients for a simplex face.}
\usage{
getHPCoefs(points)
}
\arguments{
\item{points}{The vertex points for the simplex face}
}
\value{
A vector of coefficients for the hyperplane
}
\description{
Get the hyperplane coefficients for a simplex face.
}
