% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{getTargetData}
\alias{getTargetData}
\title{Get target like data (one class inside the other)}
\usage{
getTargetData(a = 40, b = 40, dim = 2)
}
\arguments{
\item{a}{The number of class A}

\item{b}{The number of class B}

\item{dim}{The dimensionality of the data (2 or 3)}
}
\value{
A data matrix
}
\description{
Get target like data (one class inside the other)
}
