% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{getTestData}
\alias{getTestData}
\title{Get a test data set.}
\usage{
getTestData(An = 40, Bn = 20, Cn = 80, Am1 = 10, Am2 = 0, Bm1 = 0,
  Bm2 = 0, Cm1 = 15, Cm2 = 15, As1 = 5, As2 = As1, Bs1 = 5,
  Bs2 = Bs1, Cs1 = 5, Cs2 = Cs1, scatterC = T)
}
\arguments{
\item{An}{The number of class A}

\item{Bn}{The number of class B}

\item{Cn}{The number of class C}

\item{Am1}{The mean of X1 for class A}

\item{Am2}{The mean of X2 for class A}

\item{Bm1}{The mean of X1 for class B}

\item{Bm2}{The mean of X2 for class B}

\item{Cm1}{The mean of X1 for class C}

\item{Cm2}{The mean of X2 for class C}

\item{As1}{The std dev of X1 for class A}

\item{As2}{The std dev of X2 for class A}

\item{Bs1}{The std dev of X1 for class B}

\item{Bs2}{The std dev of X2 for class B}

\item{Cs1}{The std dev of X1 for class C}

\item{Cs2}{The std dev of X2 for class C}

\item{scatterC}{Whether to 'scatter' class C - see function description.}
}
\value{
A data matrix
}
\description{
Get a test data set with three classes, all drawn from normal distributions. The last 
can be randomly multiplied by 1 or -1 using scatterC.
}
