% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateKLD.R
\name{makeSampleFromDomainFunction}
\alias{makeSampleFromDomainFunction}
\title{Make sampleFromDomain function for estimating KLD}
\usage{
makeSampleFromDomainFunction(domain)
}
\arguments{
\item{domain}{A 2 rows matrix specifying domain using mins (row 1) and maxes (row 2)}
}
\value{
sampleFromDomain function
}
\description{
Make sampleFromDomain function for estimating KLD
}
\keyword{internal}
