% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat_neigh_regression.R
\name{natural_regression}
\alias{natural_regression}
\title{Natural neighbors regression}
\usage{
natural_regression(x, y, order = 1, l2 = 1e-04, verbose = 1)
}
\arguments{
\item{x}{Feature matrix}

\item{y}{Regression values}
}
\value{
A nn_regeression object with the following fields:
}
\description{
Natural neighbors regression
}
