% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSimplexes.R
\name{pdom}
\alias{pdom}
\title{Function that checks for p-dominance}
\usage{
pdom(pnts, tes, alpha)
}
\arguments{
\item{pnts}{The point(s) to test at}

\item{tes}{The tessellations object}

\item{alpha}{The test parameter}
}
\value{
Boolean vector specifying whether the target class's probability is higher than 
alpha at each point.
}
\description{
Function that checks for p-dominance. Passed to analyzeTessellation.
}
