% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotmultiple}
\alias{plotmultiple}
\title{Plot tessellations}
\usage{
plotmultiple(tes, dens)
}
\arguments{
\item{tes}{The tessellation object}

\item{dens}{Whether the density should be plotted}
}
\description{
Plot tessellations
}
\keyword{internal}
