% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotsingle}
\alias{plotsingle}
\title{Plot information from a single tessellation}
\usage{
plotsingle(tes, dens, single, triangles, tricol)
}
\arguments{
\item{tes}{The tessellation object}

\item{dens}{Whether the density should be plotted}

\item{single}{The class to plot}

\item{triangles}{The triangles to shade. Used only when not doing a density plot 
and using two dimensional data.}

\item{tricol}{The color to use when shading triangles}
}
\description{
Plot information from a single tessellation
}
\keyword{internal}
