% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plottriangle}
\alias{plottriangle}
\title{Plot a triangle}
\usage{
plottriangle(tes, tri, data, tricol)
}
\arguments{
\item{tes}{The tessellation object}

\item{tri}{The index of the triangle to plot}

\item{data}{The data points the triangle indices refer to}

\item{tricol}{The color to use when plotting}
}
\description{
Plot a triangle
}
\keyword{internal}
