% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{potential}
\alias{potential}
\title{¨Get the density potential at given points}
\usage{
potential(tes, points)
}
\arguments{
\item{tes}{The tessellations object}

\item{points}{The points}
}
\value{
Numeric matrix of density values, with rows corresponding to classes, 
and columns to points.
}
\description{
¨Get the density potential at given points
}
