% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSimplexes.R
\name{pqdom}
\alias{pqdom}
\title{Function that checks for pq-dominance}
\usage{
pqdom(pnts, tes, alpha, beta)
}
\arguments{
\item{pnts}{The point(s) to test at}

\item{tes}{The tessellations object}

\item{alpha}{One test parameter}

\item{beta}{The other test parameter}
}
\value{
Boolean specifying whether (i) the target class's probability is higher than 
alpha at the point; and (ii) whether the target class's raw density is higher than 
beta at the point.
}
\description{
Function that checks for pq-dominance. Passed to analyzeTessellation.
}
