% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict.natural_regression_mnr}
\alias{predict.natural_regression_mnr}
\title{Predict}
\usage{
\method{predict}{natural_regression_mnr}(model, X)
}
\arguments{
\item{model}{The natural neighbors regression model}

\item{X}{The feature matrix or vector to regress on}
}
\value{
Numeric giving regression estimate
}
\description{
Predict regression estimate at point.
}
