% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict.tessellations}
\alias{predict.tessellations}
\title{Predict}
\usage{
\method{predict}{tessellations}(tes, x)
}
\arguments{
\item{tes}{The tessellation object}

\item{x}{The point}
}
\value{
Integer giving index of most probable class
}
\description{
Predict (the first) most probable class at point. If probability returns NAs at a point, 
this returns NA at that point. If probabilities of classes are equal, this returns the
smaller index.
}
