% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{probability}
\alias{probability}
\title{¨Get the class probabilities at given points}
\usage{
probability(tes, points)
}
\arguments{
\item{tes}{The tessellations object}

\item{points}{The points}
}
\value{
Numeric matrix of probability values, with rows corresponding to classes, 
and columns to points.
}
\description{
¨Get the class probabilities at given points. If no class densities are non-zero at a point
then the result is NAs.
}
