% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{singledensityplot}
\alias{singledensityplot}
\title{Plot the density of a single class.}
\usage{
singledensityplot(tes, single, alpha = 1)
}
\arguments{
\item{tes}{The tessellations object}

\item{single}{The class to plot}

\item{alpha}{The alpha parameter for the density shading.}
}
\description{
Plot the density of a single class. Data coordinates must be two dimensional.
}
\keyword{internal}
