% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{singletetrameshplot}
\alias{singletetrameshplot}
\title{Plot tessellations of single class in three dimensions.}
\usage{
singletetrameshplot(tes, single)
}
\arguments{
\item{tes}{The tessellations object}

\item{single}{The class to plot}
}
\description{
Plot tessellations of single class in three dimensions.
}
\keyword{internal}
