% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{singletrimeshplot}
\alias{singletrimeshplot}
\title{Plot the tessellation of a single class in two dimensions.}
\usage{
singletrimeshplot(tes, single, triangles, tricol = single)
}
\arguments{
\item{tes}{The tessellations object}

\item{single}{The class to plot}

\item{triangles}{The triangles to shade}

\item{tricol}{The color to use when shading triangles}
}
\description{
Plot the tessellation of a single class. Data coordinates must be two dimensional.
}
\keyword{internal}
