% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellations.R
\name{tessellations}
\alias{tessellations}
\title{Create tessellations}
\usage{
tessellations(x, y = rep(1, nrow(x)), target = NA, method = 1,
  domain = NULL, outer = Inf, generative = F, dummy_boundary_points = T,
  scatter = NA, multiply_duplicates = T, l2 = 1e-04)
}
\arguments{
\item{x}{The data point coordinates}

\item{y}{The classes of the data points}

\item{target}{The first to target for polytopic analysis.  Targeting
only functions with two classes. Default is NA, 
which will target the first class if there are two classes. 
Use 0 to target no classes. A target is required to use the analyseSimplexes function on the tessellation object.}

\item{method}{Not used - should be set to 1.}

\item{domain}{The domain of the data. If not specified, the min and max values of each column are used.
This can be used to calculate the size of the polytope outside the convex hull of the different tessellations
(see outer).}

\item{outer}{The size of the polytope outside the convex hull of the different tessellations. NA will 
calculate this from the domain.}

\item{generative}{If TRUE the hypervolumes of the lifted polytopes are calculated. This
enables the tessellations to be used as a generative model.}
}
\value{
A tessellations object with the following fields:
\describe{
 \item{"tes"}{A list containing the class tessellations. This contains the fields:
   \describe{
     \item{"tri"}{A matrix giving the data points involved in the triangles of the tessellation. These
     refer to rows in the corresponding class data matrix.}
     \item{"neighbors"}{A list giving the neighbors for each triangle. If positive, these values refer to
     the triangles in the trianlge matrix. If negative, then the triangle borders the exterior polytope
     on a given edge.}
     \item{"area"}{The areas of the triangles.}
   }
 }
 \item{classdata}{Matrices giving the data coordinates for the different classes.}
 \item{x}{A matrix giving the coordinates of the complete data set.}
 \item{y}{A vector giving the classes of the complete data set.}
 \item{dim}{The number of dimensions.}
 \item{chulls}{A list of matrices the convex hulls of each tessellation/class data set.}
 \item{outer}{A vector specifying the area lying outside each tessellation.}
 \item{domain}{A matrix specifying the domain.}
 \item{classes}{A vector containing class labels (numbers if all labels are numeric).}
 \item{densities}{A vector specifying the point density estimates of the class data points.}
 \item{point_mappings}{For internal use.}
 \item{target}{The targeted class.}
 \item{critical_points_map}{For internal use.}
 \item{simplex_models}{A list of matrices, one for each class.  The nth column gives the 
 linear model associated with the nth polytope in the tessellation for that class.}
 \item{hypervolumes}{The hypervolumes associated with the polytopes of each tessellation.}
 \item{simplex_cond_probabilities}{The probability that a randomly sample from a 
 density comes from a given polytope. The normalized hypervolumes.}
}
}
\description{
Create tessellations from data
}
