% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tetrameshplot}
\alias{tetrameshplot}
\title{Plot tessellations in three dimensions.}
\usage{
tetrameshplot(tes)
}
\arguments{
\item{tes}{The tessellations object}
}
\description{
Plot tessellations in three dimensions.
}
\keyword{internal}
