% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{trimeshplot}
\alias{trimeshplot}
\title{Plot tessellations in two dimensions.}
\usage{
trimeshplot(tes)
}
\arguments{
\item{tes}{The tessellations object}
}
\description{
Plot tessellations in two dimensions.
}
\keyword{internal}
